# jara-market-backend/jaraman/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Jaramarket API",
        "description": "Jaramarket Meals Marketplace API for Customer and Vendor mobile applications",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "L5_SWAGGER_CONST_HOST",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Customer Authentication",
                    "Vendor Authentication"
                ],
                "summary": "Register a new user",
                "description": "Creates a new user account (Customer or Vendor). A verification OTP will be sent to the email address.",
                "operationId": "5b1a17e861f6e3011d01d52e8be7aaa3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "firstname",
                                    "lastname",
                                    "email",
                                    "password",
                                    "role"
                                ],
                                "properties": {
                                    "firstname": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "lastname": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Password123!"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Password123!"
                                    },
                                    "role": {
                                        "type": "string",
                                        "example": "customer",
                                        "enum": [
                                            "customer",
                                            "vendor"
                                        ]
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "08012345678"
                                    },
                                    "referral_code": {
                                        "type": "string",
                                        "example": "REF123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Success - OTP sent to email",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "An OTP has been sent to your email address."
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        },
        "/jaram/resend-otp": {
            "post": {
                "tags": [
                    "Customer Authentication",
                    "Vendor Authentication"
                ],
                "summary": "Resend verification OTP",
                "description": "Resends a new 6-digit OTP to the user's email address.",
                "operationId": "235c10572c417137720d617aa5ed35fb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "OTP Resent"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                }
            }
        },
        "/jaram/validate-otp": {
            "post": {
                "tags": [
                    "Customer Authentication",
                    "Vendor Authentication"
                ],
                "summary": "Verify email with OTP",
                "description": "Validates the 6-digit OTP sent to the user's email. On success, returns a Sanctum Bearer token for immediate login.",
                "operationId": "d7131a5386de60d5573766ac87a1bc85",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "otp"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "otp": {
                                        "type": "string",
                                        "example": "123456"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Email verified & Token generated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "OTP validated successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "1|abcdef123456..."
                                                },
                                                "user": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid or expired OTP"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Customer Authentication",
                    "Vendor Authentication"
                ],
                "summary": "Login user",
                "description": "Authenticates a user and returns a Sanctum Bearer token.",
                "operationId": "b981558f87ec3946f53132922602adbe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Password123!"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Login successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User Authenticated successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "token": {
                                                    "type": "string",
                                                    "example": "2|xyz789..."
                                                },
                                                "user": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials"
                    },
                    "403": {
                        "description": "Account not verified"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Customer Authentication",
                    "Vendor Authentication"
                ],
                "summary": "Logout user",
                "description": "Revokes the current access token.",
                "operationId": "e6cdf3fb73b369f5356a7af7bf1d7b5e",
                "responses": {
                    "200": {
                        "description": "Logged out successfully"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter token in format (Bearer <token>)",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}